Fixed broken handling of Sass changes when Grunt is running
Added grunt-sass to devDependencies in package.json so changes to Sass files in dev are handled by Grunt.
Previously when a Sass file changed, Grunt would notice the change to the file but fail to run Sass, generating
this warning: "Warning: Task "sass" not found. Use --force to continue."

The problem was caused by the yeoman template not including grunt-sass in the devDependencies.